refactor: move LoginScreen login to its own file
