+++ src/presentation/screens/login-screen/login-screen.tsx
@@ -1,63 +1,28 @@
-import React, { useState } from 'react';
-import { useForm, Controller } from 'react-hook-form';
-import { useNavigation } from '@react-navigation/native';
-import { routes } from '~/presentation/navigation/routes';
-import { makeRemoteAuthentication } from '~/main/factories/usecases';
-
-// Adapters
-import { setCurrentAccountAdapter } from '~/main/adapters';
+import React from 'react';
+import { Controller } from 'react-hook-form';
 
 // Components
 import { InputErrorMessage, TextInput } from '~/presentation/components/inputs';
 import { MainButton } from '~/presentation/components/buttons';
 import { HelperText } from 'react-native-paper';
 
+// Hooks
+import { useLoginScreenViewController } from './login-screen-view-controller';
+
 // Styles
 import { ScreenContainer } from './styles';
 
-// Types
-import { LoginFormInterface } from './types';
-import { StackNavigation } from '~/presentation/navigation/navigators/types';
-
-// Validators
-import { emailValidation } from '~/presentation/validators';
-
 export function LoginScreen() {
-  const [hasAuthenticationFailed, setHasAuthenticationFailed] = useState(false);
-  const [isLoading, setIsLoading] = useState(false);
   const {
     control,
     handleSubmit,
-    watch,
-    formState: { errors },
-  } = useForm<LoginFormInterface>({
-    defaultValues: {
-      email: '',
-      password: '',
-    },
-  });
-  const navigation = useNavigation<StackNavigation>();
-  const emailValue = watch('email');
-  const passwordValue = watch('password');
-  const hasEmailError = errors.email?.type === 'validateEmail';
-  const isMainButtonDisabled = emailValue === '' || passwordValue === '' || hasEmailError;
-
-  async function handleLogin(data: LoginFormInterface) {
-    setIsLoading(true);
-    setHasAuthenticationFailed(false);
-    const remoteAuthentication = makeRemoteAuthentication();
-
-    try {
-      const response = await remoteAuthentication.execute(data);
-
-      await setCurrentAccountAdapter(response.data);
-      setIsLoading(false);
-      navigation.navigate(routes.HomeScreen);
-    } catch (error) {
-      setIsLoading(false);
-      setHasAuthenticationFailed(true);
-    }
-  }
+    handleLogin,
+    hasAuthenticationFailed,
+    isMainButtonDisabled,
+    isLoading,
+    emailValidation,
+    hasEmailError,
+  } = useLoginScreenViewController();
 
   return (
     <ScreenContainer>
